Rename module to extend
