+++ test/extend.test.ts
@@ -1,4 +1,4 @@
-import "../src/extension";
+import "../src/extend";
 
 test("Function prototype has extension method", () => {
   expect(Function.prototype).toHaveProperty("extension");
